+++ packages/@expo/config-plugins/CHANGELOG.md
@@ -4,7 +4,9 @@
 
 ### 🛠 Breaking changes
 
-### 🎉 New features
+### 🎉 New features 
+
+- [Android] Add support `edgeToEdgeEnabled` field in app config. ([#35958](https://github.com/expo/expo/pull/35958) by [@behenate](https://github.com/behenate))
 
 ### 🐛 Bug fixes
 
@@ -25,7 +27,6 @@ _This version does not introduce any user-facing changes._
 ### 🎉 New features
 
 - Support for updates.useNativeDebug. ([#35468](https://github.com/expo/expo/pull/35468) by [@douglowder](https://github.com/douglowder))
-- [Android] Add support for `enableEdgeToEdge` app config field.
 
 ### 🐛 Bug fixes